Working Dataset.java
Lots of things changed in this class. While discussing it online and with peers at school, it was suggested to use a 2 dimensional array to step through the file and then create a method for printing the results. This is a slightly different approach than what I had originally considered. By using a for loop, I could step through the array and print just a particular element(or is it index?). The first method scans through the text file and prints out one at a time, the second method is a for loop for totaling a particular index. That is returned to the main class.